requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.9.1
oauthlib==3.2.0
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
PyPDF3==1.0.6
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.10
webencodings==0.5.1
xhtml2pdf==0.2.8
